fixed onboardig button issue
