+++ app/src/main/java/com/example/nixfit/presentation/common/AppBars.kt
@@ -2,29 +2,57 @@ package com.example.nixfit.presentation.common
 
 import android.content.res.Configuration
 import androidx.annotation.DrawableRes
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Menu
+import androidx.compose.material.icons.filled.Star
+import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.ListItem
+import androidx.compose.material3.ListItemDefaults
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.ModalDrawerSheet
+import androidx.compose.material3.ModalNavigationDrawer
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.NavigationBarItemDefaults
+import androidx.compose.material3.NavigationDrawerItem
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SearchBar
+import androidx.compose.material3.SearchBarDefaults
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.rememberDrawerState
 import androidx.compose.material3.rememberTopAppBarState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.semantics.isTraversalGroup
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.semantics.traversalIndex
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
@@ -35,6 +63,8 @@ import com.example.nixfit.presentation.fooddiary.FoodDiaryScreen
 import com.example.nixfit.presentation.navigation.Screen
 import com.example.nixfit.presentation.recipe.RecipeScreen
 import com.example.nixfit.ui.theme.NixFitTheme
+import kotlinx.coroutines.launch
+import com.example.nixfit.presentation.common.BarcodeScannerUtil
 
 // Scaffold that contains top & bottom app bars
 // Top: Hamburger icon, title, barcode scanner button
@@ -68,11 +98,106 @@ fun AppBars() {
                 backStackState?.destination?.route == Screen.Recipe.route
     }
     val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior(rememberTopAppBarState())
-
+    val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
+    val scope = rememberCoroutineScope()
+    val context = LocalContext.current
+    
     // main composable
     Scaffold(
         topBar = {
-            FoodDiaryScreen()
+            ModalNavigationDrawer(
+                drawerState = drawerState,
+                drawerContent = {
+                    ModalDrawerSheet {
+                        Text("Drawer title", modifier = Modifier.padding(16.dp))
+                        HorizontalDivider()
+                        NavigationDrawerItem(
+                            label = { Text(text = "Drawer Item") },
+                            selected = false,
+                            onClick = { /*TODO*/ }
+                        )
+                        // ...other drawer items
+                    }
+                }
+            ){
+                Box(
+                    Modifier
+                        .fillMaxSize()
+                        .semantics { isTraversalGroup = true }
+                ) {
+                    SearchBar(
+                        modifier = Modifier
+                            .align(Alignment.TopCenter)
+                            .semantics { traversalIndex = -1f },
+                        inputField = {
+                            SearchBarDefaults.InputField(
+                                query = text,
+                                onQueryChange = { text = it },
+                                onSearch = { expanded = false },
+                                expanded = expanded,
+                                onExpandedChange = { expanded = it },
+                                placeholder = { Text("Hinted search text") },
+                                leadingIcon = {
+                                    IconButton(
+                                        onClick = {
+                                            scope.launch{
+                                                drawerState.open()
+                                            }
+                                        }
+                                    ) {
+                                        Icon(
+                                            Icons.Default.Menu, contentDescription = null
+                                        )
+                                    }
+                                },
+                                trailingIcon = {
+                                               IconButton(
+                                                   onClick = {
+                                                       BarcodeScannerUtil.startBarcodeScanning(context)
+                                                   }
+                                               ){
+                                                   Icon(painterResource(R.drawable.barcode), contentDescription = "Barcode scanner")
+                                               }
+                                },
+                            )
+                        },
+                        expanded = expanded,
+                        onExpandedChange = { expanded = it },
+                    ) {
+                        repeat(4) { idx ->
+                            val resultText = "Suggestion $idx"
+                            ListItem(
+                                headlineContent = { Text(resultText) },
+                                supportingContent = { Text("Additional info") },
+                                leadingContent = { Icon(Icons.Filled.Star, contentDescription = null) },
+                                colors = ListItemDefaults.colors(containerColor = Color.Transparent),
+                                modifier = Modifier
+                                    .clickable {
+                                        text = resultText
+                                        expanded = false
+                                    }
+                                    .fillMaxWidth()
+                                    .padding(horizontal = 16.dp, vertical = 4.dp)
+                            )
+                        }
+                    }
+
+                    LazyColumn(
+                        contentPadding = PaddingValues(start = 16.dp, top = 72.dp, end = 16.dp, bottom = 16.dp),
+                        verticalArrangement = Arrangement.spacedBy(8.dp)
+                    ) {
+                        val list = List(100) { "Text $it" }
+                        items(count = list.size) {
+                            Text(
+                                text = list[it],
+                                modifier = Modifier
+                                    .fillMaxWidth()
+                                    .padding(horizontal = 16.dp),
+                            )
+                        }
+                    }
+                }
+            }
         },
         bottomBar = {
             if (isBottomNavVisible) {
