+++ app/src/main/java/com/example/nixfit/presentation/common/BarcodeScanner.kt
@@ -1,5 +1,6 @@
 package com.example.nixfit.presentation.common
 
+import android.content.Context
 import android.widget.Toast
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
@@ -14,40 +15,27 @@ import com.google.mlkit.vision.barcode.common.Barcode
 import com.google.mlkit.vision.codescanner.GmsBarcodeScannerOptions
 import com.google.mlkit.vision.codescanner.GmsBarcodeScanning
 
-@Composable
-fun BarcodeScanner(){
-    val context = LocalContext.current
-    val options = GmsBarcodeScannerOptions.Builder()
-        .setBarcodeFormats(
-            Barcode.FORMAT_EAN_8,
-            Barcode.FORMAT_EAN_13,
-            Barcode.FORMAT_UPC_A,
-            Barcode.FORMAT_UPC_E)
-        .enableAutoZoom()
-        .build()
-    val scanner = GmsBarcodeScanning.getClient(context, options)
-
-    Column{
-        FloatingActionButton(
-            onClick = {
-                scanner.startScan()
-                    .addOnSuccessListener { barcodes ->
-                        val barcodeNumber: String? = barcodes.rawValue
-                        Toast.makeText(context, barcodeNumber, Toast.LENGTH_LONG).show()
-                    }
-                    .addOnFailureListener { exception ->
-                        Toast.makeText(context, "Scan failed: ${exception.message}", Toast.LENGTH_LONG).show()
-                    }
-            },
-            modifier = Modifier.padding(16.dp)
-        ) {
-            Text(text = "Scan barcode")
-        }
+// Logic for barcode scanning
+object BarcodeScannerUtil {
+    fun startBarcodeScanning(context: Context) {
+        val options = GmsBarcodeScannerOptions.Builder()
+            .setBarcodeFormats(
+                Barcode.FORMAT_EAN_8,
+                Barcode.FORMAT_EAN_13,
+                Barcode.FORMAT_UPC_A,
+                Barcode.FORMAT_UPC_E
+            )
+            .enableAutoZoom()
+            .build()
+        val scanner = GmsBarcodeScanning.getClient(context, options)
+        scanner.startScan()
+            .addOnSuccessListener { barcodes ->
+                val barcodeNumber: String? = barcodes.rawValue
+                Toast.makeText(context, barcodeNumber, Toast.LENGTH_LONG).show()
+            }
+            .addOnFailureListener { exception ->
+                Toast.makeText(context, "Scan failed: ${exception.message}", Toast.LENGTH_LONG)
+                    .show()
+            }
     }
-}
-
-@Preview
-@Composable
-fun BarcodeScannerPreview(){
-    BarcodeScanner()
 }
