+++ app/src/main/java/com/example/nixfit/presentation/fooddiary/FoodDiaryScreen.kt
@@ -46,95 +46,7 @@ import kotlinx.coroutines.launch
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun FoodDiaryScreen () {
-    var text by rememberSaveable { mutableStateOf("") }
-    var expanded by rememberSaveable { mutableStateOf(false) }
-    val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
-    val scope = rememberCoroutineScope()
-    ModalNavigationDrawer(
-        drawerState = drawerState,
-        drawerContent = {
-            ModalDrawerSheet {
-                Text("Drawer title", modifier = Modifier.padding(16.dp))
-                HorizontalDivider()
-                NavigationDrawerItem(
-                    label = { Text(text = "Drawer Item") },
-                    selected = false,
-                    onClick = { /*TODO*/ }
-                )
-                // ...other drawer items
-            }
-        }
-    ){
-        Box(
-            Modifier
-                .fillMaxSize()
-                .semantics { isTraversalGroup = true }
-        ) {
-            SearchBar(
-                modifier = Modifier
-                    .align(Alignment.TopCenter)
-                    .semantics { traversalIndex = -1f },
-                inputField = {
-                    SearchBarDefaults.InputField(
-                        query = text,
-                        onQueryChange = { text = it },
-                        onSearch = { expanded = false },
-                        expanded = expanded,
-                        onExpandedChange = { expanded = it },
-                        placeholder = { Text("Hinted search text") },
-                        leadingIcon = {
-                            IconButton(
-                                onClick = {
-                                    scope.launch{
-                                        drawerState.open()
-                                    }
-                                }
-                            ) {
-                                Icon(
-                                    Icons.Default.Menu, contentDescription = null
-                                )
-                            }
-                        },
-                        trailingIcon = { Icon(painterResource(R.drawable.barcode), contentDescription = null) },
-                    )
-                },
-                expanded = expanded,
-                onExpandedChange = { expanded = it },
-            ) {
-                repeat(4) { idx ->
-                    val resultText = "Suggestion $idx"
-                    ListItem(
-                        headlineContent = { Text(resultText) },
-                        supportingContent = { Text("Additional info") },
-                        leadingContent = { Icon(Icons.Filled.Star, contentDescription = null) },
-                        colors = ListItemDefaults.colors(containerColor = Color.Transparent),
-                        modifier = Modifier
-                            .clickable {
-                                text = resultText
-                                expanded = false
-                            }
-                            .fillMaxWidth()
-                            .padding(horizontal = 16.dp, vertical = 4.dp)
-                    )
-                }
-            }
 
-            LazyColumn(
-                contentPadding = PaddingValues(start = 16.dp, top = 72.dp, end = 16.dp, bottom = 16.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                val list = List(100) { "Text $it" }
-                items(count = list.size) {
-                    Text(
-                        text = list[it],
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(horizontal = 16.dp),
-                    )
-                }
-            }
-        }
-    }
 }
