+++ app/src/main/java/com/example/nixfit/presentation/onboarding/OnBoardingEvent.kt
@@ -1,5 +1,5 @@
 package com.example.nixfit.presentation.onboarding
 
 sealed class OnBoardingEvent {
-    object SaveAppEntry : OnBoardingEvent()
+    data object SaveAppEntry : OnBoardingEvent()
 }
