+++ app/src/main/java/com/example/nixfit/presentation/onboarding/OnBoardingViewModel.kt
@@ -2,20 +2,23 @@ package com.example.nixfit.presentation.onboarding
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.nixfit.domain.usecases.appentry.SaveAppEntry
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class OnBoardingViewModel @Inject constructor() : ViewModel(){
+class OnBoardingViewModel @Inject constructor(
+    private val saveAppEntry: SaveAppEntry
+) : ViewModel(){
     fun onEvent(event: OnBoardingEvent){
         when(event){
             is OnBoardingEvent.SaveAppEntry -> {
-                saveAppEntry()
+                saveUserEntry()
             }
         }
     }
-    private fun saveAppEntry(){
+    private fun saveUserEntry(){
         viewModelScope.launch {
             saveAppEntry()
         }
